Rework how MIG device placements are tracked and presented
We no longer show the "available" count of GPUs nor a count of MIG devices in
the CRD.  Instead, we only show the possible placements on a GPU-by-GPU basis
for a given profile. From this and the set of already allocated devices on a
node, the controller has all of the information it needs to determine if there
is enough space to allocate either a GPU or MIG device.

Signed-off-by: Kevin Klues <kklues@nvidia.com>