Add helm option to set driver root
This change adds an nvidiaDriverRoot helm option that allows
the root of the NVIDIA GPU Driver to be specified. If this is
not specified, then a driver installed directly on the node is
assumed.

This change also changes the mount location of the driver root
in the contianer to /driver-root instead of /run/nvidia/driver
to align with the paths used by the device plugin. This should
also serve to reduce confusion between the two paths -- especially
in the case where the driver container is used.

Signed-off-by: Evan Lezar <elezar@nvidia.com>