Include common edits in claim spec
This change includes common edits (e.g. driver library mounts)
in the specifications created for each claim. The edits for the
MPS server are also moved to the common edits section.

This means that a spec file for a claim is self-contained, and
it is not required to include a "common" dummy device.

Signed-off-by: Evan Lezar <elezar@nvidia.com>