Add logic to discover libnvidia-ml.so and nvidia-smi
This change adds logic to discover the full paths to libnvidia-ml.so.1 and
nvidia-smi in the driver root. These paths are used to load the correct
library when interacting with the NVML package and also when running
nvidia-smi.

This requires changes to go-nvlib (with go-nvml as a transitive dependency)
that allows the library path to be explicitly set.

Signed-off-by: Evan Lezar <elezar@nvidia.com>