Switch to pipes for controlling MPS
This change switches from environment variables -- which are intended as
client-side controls -- to piping values to the MPS control daemon.

This also allows for a file-bases approach to be used for the readiness
check, with a file created as soon as the MPS control daemon is ready and the
relevant values configured.

Signed-off-by: Evan Lezar <elezar@nvidia.com>