Let kublet plugin run privileged on OpenShift
On OpenShift, the kublet plugin container won't run without applying the right
security settings that let it run privileged. We solve this by binding the privileged
security context constraints (SCC) role to the DRA driver's service account.

This is done only when the target system has the OpenShift SCC capability in order
to minimise the impact on non-OpenShift platforms.

Signed-off-by: Vitaliy Emporopulo <vemporop@redhat.com>