Update logic to set environment for calls out to nvidia-smi
Previously, we were setting the envvars needed for nvidia-smi in the plugin's
environment itself. This caused errors, however, when shelling out to other
binaries that didn't need these envvars set.

This change pushes the setting of the envvars into the actual exec call for
nvidia-smi instead of setting them in the plugin's environment itself.

Closes #106

Signed-off-by: Kevin Klues <kklues@nvidia.com>